wip: organize hub package by top level steps and dependent components
The hub package is starting to get large and difficult to understand how things relate. This is an initial attempt
to start breaking up the hub package into sub components.
