+++ hub/cluster/manage.go
@@ -1,0 +1,59 @@
+package cluster
+
+import (
+	"fmt"
+	"os/exec"
+
+	"github.com/greenplum-db/gpupgrade/step"
+	"github.com/greenplum-db/gpupgrade/utils"
+)
+
+func StopCluster(stream step.OutStreams, cluster *utils.Cluster) error {
+	return startStopCluster(stream, cluster, true)
+}
+func StartCluster(stream step.OutStreams, cluster *utils.Cluster) error {
+	return startStopCluster(stream, cluster, false)
+}
+
+var isPostmasterRunningCmd = exec.Command
+var startStopClusterCmd = exec.Command
+
+func IsPostmasterRunning(stream step.OutStreams, cluster *utils.Cluster) error {
+	cmd := isPostmasterRunningCmd("bash", "-c",
+		fmt.Sprintf("pgrep -F %s/postmaster.pid",
+			cluster.MasterDataDir(),
+		))
+
+	cmd.Stdout = stream.Stdout()
+	cmd.Stderr = stream.Stderr()
+
+	return cmd.Run()
+}
+
+func startStopCluster(stream step.OutStreams, cluster *utils.Cluster, stop bool) error {
+	// TODO: why can't we call IsPostmasterRunning for the !stop case?  If we do, we get this on the pipeline:
+	// Usage: pgrep [-flvx] [-d DELIM] [-n|-o] [-P PPIDLIST] [-g PGRPLIST] [-s SIDLIST]
+	// [-u EUIDLIST] [-U UIDLIST] [-G GIDLIST] [-t TERMLIST] [PATTERN]
+	//  pgrep: pidfile not valid
+	// TODO: should we actually return an error if we try to gpstop an already stopped cluster?
+	cmdName := "gpstart"
+	if stop {
+		cmdName = "gpstop"
+		err := IsPostmasterRunning(stream, cluster)
+		if err != nil {
+			return err
+		}
+	}
+
+	cmd := startStopClusterCmd("bash", "-c",
+		fmt.Sprintf("source %[1]s/../greenplum_path.sh && %[1]s/%[2]s -a -d %[3]s",
+			cluster.BinDir,
+			cmdName,
+			cluster.MasterDataDir(),
+		))
+
+	cmd.Stdout = stream.Stdout()
+	cmd.Stderr = stream.Stderr()
+
+	return cmd.Run()
+}
