+++ hub/upgrade_master_test.go
@@ -273,17 +273,3 @@ func (b *bufferedStreams) Stdout() io.Writer {
 func (b *bufferedStreams) Stderr() io.Writer {
 	return &b.stderr
 }
-
-// failingStreams is an implementation of OutStreams for which every call to a
-// stream's Write() method will fail with the given error.
-type failingStreams struct {
-	err error
-}
-
-func (f failingStreams) Stdout() io.Writer {
-	return &failingWriter{f.err}
-}
-
-func (f failingStreams) Stderr() io.Writer {
-	return &failingWriter{f.err}
-}
